# adservice/src/main/resources/api/openapi.yaml
openapi: 3.0.0
servers:
  - url: 'http://www.ebayadservices.com/marketingtracking/v1'
    description: eBay Adservices
info:
  description: eBay Adservices
  version: 1.0.0
  title: ebay-adservices
paths:
...
  /privacy-sandbox-attestations.json:
    get:
      summary: Privacy Sandbox Attestations
      description: Privacy Sandbox Attestations
      tags:
        - attestations
      operationId: attestations
      responses:
        '200':
          description: Success
        '400':
          description: Rejected
        '500':
          description: The call was unsuccessful due to exceptions on the Server
...